squasher_inspector/pyproject.toml
[tool.poetry]
name = "squasher_inspector"
version = "0.1.0"
description = ""
authors = ["wappon28dev <wappon_28_dev@ruru.be>"]
readme = "README.md"
packages = [{ include = "squasher_inspector" }]

[tool.poetry.dependencies]
python = "3.11.2"
black = "^23.7.0"
bardapi = "0.1.33"
dataclasses-json = "^0.6.1"
python-dotenv = "^1.0.0"
matplotlib = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
